refactor(selector4nix): move `NarStreamProvider` to `NarFile` aggregate
