+++ src/api/handlers/nar.rs
@@ -8,8 +8,8 @@ use futures::StreamExt;
 use crate::api::state::AppContext;
 use crate::application::AppError;
 use crate::domain::nar_file::model::NarFileKey;
+use crate::domain::nar_file::port::NarStreamData;
 use crate::domain::nar_info::model::NarFileName;
-use crate::domain::nar_info::port::NarStreamData;
 
 pub async fn get_nar(
     State(ctx): State<Arc<AppContext>>,
