+++ src/domain/nar_file/mod.rs
@@ -1,2 +1,3 @@
 pub mod model;
+pub mod port;
 pub mod service;
