+++ src/domain/nar_file/port/mod.rs
@@ -1,0 +1,3 @@
+mod nar_stream_provider;
+
+pub use nar_stream_provider::{NarStreamData, NarStreamHeaders, NarStreamProvider};
