+++ src/domain/nar_file/port/nar_stream_provider.rs
