+++ src/domain/nar_info/port/mod.rs
@@ -1,5 +1,3 @@
 mod nar_info_provider;
-mod nar_stream_provider;
 
 pub use nar_info_provider::{NarInfoProvider, NarInfoQueryData, QueryNarInfoError, error_ctx};
-pub use nar_stream_provider::{NarStreamData, NarStreamHeaders, NarStreamProvider};
